WebComponentsの登録処理を切り出し
