+++ src/index.js
@@ -4,6 +4,7 @@ import { process } from './data.js';
 import { createArticle, createPostlist, createTagged } from './model.js';
 import { renderHTML } from './render.js';
 import { realTimeSearch } from './search.js';
+import { registerComponents } from './component.js';
 
   // 表示調整用
   const jsonPath = './data.json';
@@ -36,40 +37,6 @@ import { realTimeSearch } from './search.js';
     document.querySelector('.el_search_form').addEventListener('input', () => {
         realTimeSearch(data);
     });
-
-    // ブログカード
-    class BlogCard extends HTMLElement {
-      constructor() {
-        super();
-        this.id = this.getAttribute('id');
-        this.i = data.length - this.id;
-        this.hashtags = data[this.i].tags
-            .map((eachTag) => `<li>#${eachTag}</li>`)
-            .join('');
-        this.innerHTML = `
-        <div class="bl_blogcard">
-            <a href="?id=${this.id}">
-            <header class="bl_blogcard_header">
-                <div class="bl_blogcard_icon"></div>
-                <div class="bl_blogcard_logo">placet experiri</span>
-                <span class="bl_blogcard_suffix"> :: ${this.id}</span>
-            </header>
-            <div class="bl_blogcard_title">${data[this.i].title}</div>
-            <p class="bl_blogcard_text">${data[this.i].plainText.substr(0, 56)}…</p>
-            <footer class="bl_blogcard_footer">
-                <span class="bl_blogcard_time">${dayjs(data[this.i].date).format("YYYY-MM-DD")}</span>
-                <ul class="bl_blogcard_tags">
-                ${this.hashtags}
-                </ul>
-            </footer>
-            </a>
-        </div>`;
-      }
-    }
-
-    // レンダリングと結びつける
-    // 第一引数にHTML上のカスタムエレメント、第二引数にJS上のクラス名
-    window.customElements.define('blog-card', BlogCard);
   }
 
 // JSONデータを取得して実行
@@ -77,6 +44,7 @@ const index = async () => {
   const response = await fetch(jsonPath);
   const data = await response.json();
   app(data);
+  registerComponents(data);
 }
 
 index();
